Enabled cluster tests in and split test_placement_group_n.py
