+++ python/ray/tests/BUILD
@@ -109,6 +109,11 @@ py_test_module_list(
     "test_placement_group_3.py",
     "test_placement_group_4.py",
     "test_placement_group_5.py",
+    "test_placement_group_6.py",
+    "test_placement_group_7.py",
+    "test_placement_group_8.py",
+    "test_placement_group_9.py",
+    "test_placement_group_10.py",
     "test_ray_init.py",
     "test_reference_counting.py",
     "test_resource_demand_scheduler.py",
